RSVP SQL Case Study
RSVP Movies is an Indian film production company that has produced many super-hit movies. They have usually released movies for the Indian audience but for their next project, they are planning to release a movie for a global audience in 2022.

The production company wants to plan its every move analytically based on data and has approached you for help with this new project. You have been provided with the data on the movies that have been released in the past three years. You have to analyze the data set and draw meaningful insights that can help them start their new project.